move PlayerGui to separate package
