+++ app/src/main/java/com/github/dawidd6/andttt/fragments/BaseGameFragment.java
@@ -1,6 +1,5 @@
 package com.github.dawidd6.andttt.fragments;
 
-import android.app.Fragment;
 import android.content.pm.ActivityInfo;
 import android.graphics.Bitmap;
 import android.graphics.Color;
@@ -22,6 +21,7 @@ import com.github.dawidd6.andttt.drawings.DrawCross;
 import com.github.dawidd6.andttt.drawings.DrawLine;
 import com.github.dawidd6.andttt.game.Game;
 import com.github.dawidd6.andttt.game.Player;
+import com.github.dawidd6.andttt.gui.PlayerGui;
 import com.github.dawidd6.andttt.proto.Symbol;
 
 public abstract class BaseGameFragment extends BaseFragment {
